safari-600.7-branch/LayoutTests/fast/dom/serialize-nodes-expected.txt

ALERT: foo:name is one and should be one
ALERT: bar:name is two and should be two
ALERT: name is three and should be three
ALERT: node is child and should be child
ALERT: <foo:root xmlns:foo="urn:foo-ns" foo:type="test"><child foo:name="one" bar:name="two" xmlns:bar="urn:bar-ns" name="three"/></foo:root>
<div xmlns="http://www.w3.org/1999/xhtml" id="input">
    <div>
        <foo:node xmlns:foo="http://foo.com" xmlns="http://baz.com" foo:name="foo_name" bar:name="bar_name" xmlns:bar="http://bar.com">
            <node foo:name="foo_name">
                <bar:node xmlns:bar="http://bar2.com"/>
            </node>
        </foo:node>
    </div>
    <bar:node xmlns:bar="http://bar.com">
        <br />
    </bar:node>
</div>
